fix: move libp2p logger into foundation - use pino directly
