+++ yarn-project/foundation/src/log/index.ts
@@ -3,3 +3,4 @@ export * from './debug.js';
 export * from './pino-logger.js';
 export * from './log_history.js';
 export * from './log_fn.js';
+export * from './libp2p_logger.js';
